chore: Perf optimization
- split up fixture details into many pages
