+++ app/routes/_with-navbar+/fixtures+/$id+/_index.tsx
@@ -1,0 +1,10 @@
+import { LoaderFunctionArgs, redirect } from "@remix-run/node";
+import invariant from "tiny-invariant";
+
+export const loader = async ({ params }: LoaderFunctionArgs) => {
+  const { id } = params;
+
+  invariant(typeof id === "string", "Fixture id must be string");
+
+  return redirect(`/fixtures/${id}/events`);
+};
